Version 2 (lots of changes)
This new version works a bit differently in that it will take use the grouped z function at any interval reasonably smartly. The downside is that it will cut off the frames at the end of the recording. It will report how many frames were removed at the end of the batch.