new LUT
This is a lut I got from a paper which suggested that most of the other LUTS don't scale the perceived change with the actual change, e.g. our brains don't scale brightness/color changes linearly. This is one of the many LUTs they provided which do scale linearly with the visual perception.